# server/swagger.yaml
openapi: 3.0.3
info:
  title: AGROHUB API
  version: 1.0.0
  contact: {}
servers:
  - url: '{{auth-uri}}'
  - url: '{{user-uri}}'
  - url: '{{org}}'
  - url: '{{budget}}'
  - url: '{{sub}}'
  - url: '{{transaction}}'
  - url: '{{org-user}}'
  - url: '{{product}}'
paths:
...
  /65a9192851ba18d53e60e781:
    get:
      summary: Get Single user
      description: Get Single user
      operationId: getSingleUser
      parameters:
        - name: ''
          in: header
          schema:
            type: string
            example: ''
      responses:
        '200':
          description: ''
...
  /672d40db7cb266d455d2490f:
    put:
      summary: Update Transaction
      description: Update Transaction
      operationId: updateTransaction
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  example: 15000
                budget:
                  type: string
                  example: 6720a6009543d348e43f9663
                category:
                  type: string
                  example: expense
                description:
                  type: string
                  example: Sale of agricultural products
                organization:
                  type: string
                  example: 67207bf4bf207c0ed1f7bbc3
                receipt:
                  type: string
                  example: default-recept.png
                title:
                  type: string
                  example: farm inputs expess for 2024
                transaction_date:
                  type: string
                  example: '2023-10-01'
            examples:
              Update Transaction:
                value:
                  amount: 15000
                  budget: 6720a6009543d348e43f9663
                  category: expense
                  description: Sale of agricultural products
                  organization: 67207bf4bf207c0ed1f7bbc3
                  receipt: default-recept.png
                  title: farm inputs expess for 2024
                  transaction_date: '2023-10-01'
      responses:
        '200':
          description: ''
...